🚚 move: SearchBar device 폴더 분리
