+++ app/src/components/elements/SearchBar/Mobile/MobileSearchInput.tsx
