+++ app/src/components/elements/SearchBar/Mobile/index.tsx
