+++ app/src/components/elements/SearchBar/TabletAndAbove/TabletAndAboveSearchInput.tsx
