+++ app/src/components/elements/SearchBar/TabletAndAbove/TabletAndAboveSearchResult.tsx
@@ -1,8 +1,8 @@
 import { ProjectPreview, UserPreview } from '@/__generated__/graphql';
 import { VStack } from '@components/common';
 import styled from '@emotion/styled';
-import { ProjectSearchList } from './ProjectSearchList';
-import { UserSearchList } from './UserSearchList';
+import { ProjectSearchList } from '../shared/ProjectSearchList';
+import { UserSearchList } from '../shared/UserSearchList';
 
 type TabletAndAboveSearchResultProps = {
   users: UserPreview[];
