+++ app/src/components/elements/SearchBar/shared/ProjectSearchItem.tsx
