+++ app/src/components/elements/SearchBar/shared/UserSearchItem.tsx
